ADD: firebase data redux로 저장
