+++ src/redux/store.ts
@@ -1,9 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
+import cartReducer from "./cartSlice";
 import productsReducer from "./productsSlice";
 
 export const store = configureStore({
   reducer: {
     products: productsReducer,
+    cart: cartReducer,
   },
 });
